refactor: split document context in sub-modules
