--- src/renderer/src/app-state/current-document/context.tsx
@@ -1,519 +1,0 @@
-import * as Effect from 'effect/Effect';
-import {
-  createContext,
-  useCallback,
-  useContext,
-  useEffect,
-  useRef,
-  useState,
-} from 'react';
-import { useMatch, useNavigate, useParams } from 'react-router';
-
-import {
-  isValidProjectId,
-  processDocumentChange,
-  type ProjectId,
-  type ProjectStore,
-  urlEncodeProjectId,
-} from '../../../../modules/domain/project';
-import {
-  isEmpty,
-  type RichTextDocument,
-  type VersionedDocument,
-} from '../../../../modules/domain/rich-text';
-import { RepresentationTransformContext } from '../../../../modules/domain/rich-text/react/representation-transform-context';
-import {
-  type ArtifactId,
-  type Change,
-  type ChangeId,
-  changeIdsAreSame,
-  type ChangeWithUrlInfo,
-  type Commit,
-  urlEncodeArtifactId,
-  urlEncodeChangeId,
-  urlEncodeChangeIdForChange,
-} from '../../../../modules/infrastructure/version-control';
-import { FunctionalityConfigContext } from '../../../../modules/personalization/browser';
-import { useCurrentChangeId } from '../../hooks/use-current-change-id';
-import { useCurrentDocumentId } from '../../hooks/use-current-document-id';
-import { usePulledUpstreamChanges } from '../../hooks/use-pulled-upstream-changes';
-import { InfrastructureAdaptersContext, ProjectContext } from '../';
-
-export type CurrentDocumentContextType = {
-  versionedDocumentId: ArtifactId | null;
-  versionedDocument: VersionedDocument | null;
-  onDocumentContentChange: (doc: RichTextDocument) => Promise<void>;
-  loadingHistory: boolean;
-  versionedDocumentHistory: ChangeWithUrlInfo[];
-  canCommit: boolean;
-  reloadDocumentHistory: () => Promise<void>;
-  onRestoreCommit: (args: { message: string; commit: Commit }) => Promise<void>;
-  onDiscardChanges: () => Promise<void>;
-  commitToRestore: Commit | null;
-  isRestoreCommitDialogOpen: boolean;
-  isDiscardChangesDialogOpen: boolean;
-  onOpenRestoreCommitDialog: (commit: Commit) => void;
-  onCloseRestoreCommitDialog: () => void;
-  onOpenDiscardChangesDialog: () => void;
-  onCloseDiscardChangesDialog: () => void;
-  selectedCommitIndex: number | null;
-  onSelectChange: (commitId: ChangeId) => void;
-};
-
-export const CurrentDocumentContext = createContext<CurrentDocumentContextType>(
-  {
-    versionedDocumentId: null,
-    versionedDocument: null,
-    onDocumentContentChange: async () => {},
-    loadingHistory: false,
-    versionedDocumentHistory: [],
-    canCommit: false,
-    reloadDocumentHistory: async () => {},
-    onRestoreCommit: async () => {},
-    onDiscardChanges: async () => {},
-    isRestoreCommitDialogOpen: false,
-    isDiscardChangesDialogOpen: false,
-    commitToRestore: null,
-    onOpenRestoreCommitDialog: () => {},
-    onCloseRestoreCommitDialog: () => {},
-    onOpenDiscardChangesDialog: () => {},
-    onCloseDiscardChangesDialog: () => {},
-    selectedCommitIndex: null,
-    onSelectChange: () => {},
-  }
-);
-
-export const CurrentDocumentProvider = ({
-  children,
-}: {
-  children: React.ReactNode;
-}) => {
-  const { projectStore } = useContext(InfrastructureAdaptersContext);
-  const [versionedDocument, setVersionedDocument] =
-    useState<VersionedDocument | null>(null);
-  const { projectId: projectIdParam } = useParams();
-  const changeId = useCurrentChangeId();
-  const documentId = useCurrentDocumentId();
-  const [loadingHistory, setLoadingHistory] = useState<boolean>(false);
-  const [versionedDocumentHistory, setVersionedDocumentHistory] = useState<
-    ChangeWithUrlInfo[]
-  >([]);
-  const [lastCommit, setLastCommit] = useState<Commit | null>(null);
-  const [canCommit, setCanCommit] = useState(false);
-  const [isRestoreCommitDialogOpen, setIsRestoreCommitDialogOpen] =
-    useState<boolean>(false);
-  const [isDiscardChangesDialogOpen, setIsDiscardChangesDialogOpen] =
-    useState<boolean>(false);
-  const [commitToRestore, setCommitToRestore] = useState<Commit | null>(null);
-  const [selectedCommitIndex, setSelectedCommitIndex] = useState<number | null>(
-    null
-  );
-  const { showDiffInHistoryView } = useContext(FunctionalityConfigContext);
-  const navigate = useNavigate();
-  const { restoreDocumentChanges } = useContext(ProjectContext);
-  const { adapter: representationTransformAdapter } = useContext(
-    RepresentationTransformContext
-  );
-  const prevProjectId = useRef(projectIdParam);
-  const prevDocumentId = useRef(documentId);
-  const { pulledUpstreamChanges, resetPulledUpstreamChanges } =
-    usePulledUpstreamChanges();
-  const [documentNeedsReload, setDocumentNeedsReload] = useState(false);
-  const documentRouteMatch = useMatch(
-    '/projects/:projectId/artifacts/:artifactId'
-  );
-  const documentChangeSubRouteMatch = useMatch(
-    '/projects/:projectId/artifacts/:artifactId/changes/:changeId'
-  );
-
-  useEffect(() => {
-    const projectOrDocumentHasChanged =
-      prevProjectId.current !== projectIdParam ||
-      prevDocumentId.current !== documentId;
-
-    const returningToSelectedDocumentEditMode =
-      prevDocumentId.current === documentId &&
-      prevProjectId.current === projectIdParam &&
-      !changeId;
-
-    const updateDocumentHandleAndSelectedFile = async ({
-      projectStore,
-    }: {
-      projectStore: ProjectStore;
-    }) => {
-      if (!documentId) {
-        setVersionedDocument(null);
-        resetPulledUpstreamChanges();
-        setDocumentNeedsReload(false);
-      } else {
-        setVersionedDocument(null);
-
-        const { artifact: document } = await Effect.runPromise(
-          projectStore.findDocumentById({
-            projectId: projectIdParam as ProjectId,
-            documentId,
-          })
-        );
-
-        setVersionedDocument(document);
-        setLoadingHistory(true);
-        setDocumentNeedsReload(false);
-
-        resetPulledUpstreamChanges();
-
-        prevProjectId.current = projectIdParam;
-        prevDocumentId.current = documentId;
-      }
-    };
-
-    if (
-      projectStore &&
-      projectIdParam &&
-      (projectOrDocumentHasChanged ||
-        returningToSelectedDocumentEditMode ||
-        pulledUpstreamChanges ||
-        documentNeedsReload)
-    ) {
-      updateDocumentHandleAndSelectedFile({ projectStore });
-    }
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [
-    documentId,
-    projectIdParam,
-    changeId,
-    projectStore,
-    pulledUpstreamChanges,
-    documentNeedsReload,
-  ]);
-
-  const checkIfContentChangedFromLastCommit =
-    (projectStore: ProjectStore) =>
-    async (
-      documentId: ArtifactId,
-      latestChangeId: ChangeId,
-      lastCommitId: ChangeId
-    ) => {
-      if (!changeIdsAreSame(latestChangeId, lastCommitId)) {
-        const isContentSame = await Effect.runPromise(
-          projectStore.isContentSameAtChanges({
-            projectId: projectIdParam as ProjectId,
-            documentId,
-            change1: latestChangeId,
-            change2: lastCommitId,
-          })
-        );
-
-        if (!isContentSame) {
-          setCanCommit(true);
-        } else {
-          setCanCommit(false);
-        }
-      } else {
-        setCanCommit(false);
-      }
-    };
-
-  const checkIfCanCommit =
-    (projectStore: ProjectStore) =>
-    async ({
-      docId,
-      doc,
-      latestChangeId,
-      lastCommitId,
-    }: {
-      docId: ArtifactId;
-      doc: VersionedDocument;
-      latestChangeId: ChangeId;
-      lastCommitId?: ChangeId;
-    }) => {
-      if (lastCommitId) {
-        return checkIfContentChangedFromLastCommit(projectStore)(
-          docId,
-          latestChangeId,
-          lastCommitId
-        );
-      } else {
-        if (!isEmpty(doc)) {
-          setCanCommit(true);
-        } else {
-          setCanCommit(false);
-        }
-      }
-    };
-
-  const loadHistory =
-    (projectStore: ProjectStore) =>
-    async ({ docId }: { docId: ArtifactId }) => {
-      const historyInfo = await Effect.runPromise(
-        projectStore.getDocumentHistory({
-          projectId: projectIdParam as ProjectId,
-          documentId: docId,
-        })
-      );
-
-      const historyWithURLInfo = historyInfo.history.map((commit) => ({
-        ...commit,
-        urlEncodedChangeId: urlEncodeChangeIdForChange(commit),
-      }));
-
-      setVersionedDocumentHistory(historyWithURLInfo);
-      setLastCommit(historyInfo.lastCommit);
-      setLoadingHistory(false);
-      await checkIfCanCommit(projectStore)({
-        docId,
-        doc: historyInfo.current,
-        latestChangeId: historyInfo.latestChange.id,
-        lastCommitId: historyInfo.lastCommit?.id,
-      });
-
-      return historyWithURLInfo;
-    };
-
-  useEffect(() => {
-    if (projectStore && documentId && versionedDocument) {
-      loadHistory(projectStore)({
-        docId: documentId,
-      });
-    }
-  }, [versionedDocument]);
-
-  useEffect(() => {
-    if (versionedDocumentHistory.length > 0 && changeId) {
-      const selectedCommitIndex = findSelectedCommitIndex({
-        changeId,
-        history: versionedDocumentHistory,
-      });
-
-      setSelectedCommitIndex(
-        selectedCommitIndex === -1 ? null : selectedCommitIndex
-      );
-    }
-  }, [versionedDocumentHistory, changeId]);
-
-  const reloadDocumentHistory = useCallback(async () => {
-    if (!projectStore || !versionedDocument || !documentId) return;
-    await loadHistory(projectStore)({
-      docId: documentId,
-    });
-  }, [projectStore, versionedDocument, documentId]);
-
-  const handleRestoreCommit = useCallback(
-    async ({ message, commit }: { message: string; commit: Commit }) => {
-      if (!documentId || !versionedDocument || !projectStore) {
-        throw new Error(
-          'Cannot restore commit. Either the document or the project store is not initialized yet.'
-        );
-      }
-
-      const restoreCommitId = await restoreDocumentChanges({
-        documentId,
-        commit,
-        message,
-      });
-
-      const newHistory = await loadHistory(projectStore)({
-        docId: documentId,
-      });
-
-      setIsRestoreCommitDialogOpen(false);
-      setCanCommit(false);
-      handleSelectChange(restoreCommitId, newHistory);
-    },
-    [documentId, versionedDocument, projectStore, restoreDocumentChanges]
-  );
-
-  const handleDiscardChanges = useCallback(async () => {
-    if (!documentId || !versionedDocument || !projectStore) {
-      throw new Error(
-        'Cannot discard changes. Either the document or the project store is not initialized yet.'
-      );
-    }
-
-    await Effect.runPromise(
-      projectStore.discardUncommittedChanges({
-        projectId: projectIdParam as ProjectId,
-        documentId,
-      })
-    );
-
-    const newHistory = await loadHistory(projectStore)({
-      docId: documentId,
-    });
-
-    if (documentChangeSubRouteMatch) {
-      const [lastCommit] = newHistory;
-      handleSelectChange(lastCommit.id, newHistory);
-    } else if (documentRouteMatch) {
-      setDocumentNeedsReload(true);
-    }
-
-    setIsDiscardChangesDialogOpen(false);
-    setCanCommit(false);
-  }, [
-    documentId,
-    versionedDocument,
-    projectStore,
-    documentChangeSubRouteMatch,
-    documentRouteMatch,
-  ]);
-
-  const handleOpenRestoreCommitDialog = useCallback((commit: Commit) => {
-    setIsRestoreCommitDialogOpen(true);
-    setCommitToRestore(commit);
-  }, []);
-
-  const handleCloseRestoreCommitDialog = useCallback(() => {
-    setIsRestoreCommitDialogOpen(false);
-    setCommitToRestore(null);
-  }, []);
-
-  const handleOpenDiscardChangesDialog = useCallback(() => {
-    setIsDiscardChangesDialogOpen(true);
-  }, []);
-
-  const handleCloseDiscardChangesDialog = useCallback(() => {
-    setIsDiscardChangesDialogOpen(false);
-  }, []);
-
-  const findSelectedCommitIndex = ({
-    changeId,
-    history,
-  }: {
-    changeId: ChangeId;
-    history: ChangeWithUrlInfo[];
-  }) => history.findIndex((commit) => changeIdsAreSame(commit.id, changeId));
-
-  const selectChange = ({
-    projectId,
-    documentId,
-    history,
-    changeId,
-    showDiffInHistoryView,
-  }: {
-    projectId: ProjectId;
-    documentId: ArtifactId;
-    history: ChangeWithUrlInfo[];
-    changeId: ChangeId;
-    showDiffInHistoryView: boolean;
-  }) => {
-    const isInitialChange = (index: number, changes: Change[]) =>
-      index === changes.length - 1;
-
-    const selectedCommitIndex = findSelectedCommitIndex({ changeId, history });
-
-    const isFirstCommit = isInitialChange(selectedCommitIndex, history);
-
-    const diffCommit = isFirstCommit ? null : history[selectedCommitIndex + 1];
-
-    let newUrl = `/projects/${urlEncodeProjectId(projectId)}/artifacts/${urlEncodeArtifactId(documentId)}/changes/${urlEncodeChangeId(changeId)}`;
-    if (diffCommit) {
-      const diffChangeURLEncodedId = urlEncodeChangeIdForChange(diffCommit);
-      newUrl += `?diffWith=${diffChangeURLEncodedId}`;
-    }
-
-    if (showDiffInHistoryView && diffCommit) {
-      newUrl += `&showDiff=true`;
-    }
-
-    navigate(newUrl);
-  };
-
-  const handleSelectChange = useCallback(
-    (changeId: ChangeId, history?: ChangeWithUrlInfo[]) => {
-      if (!projectIdParam || !isValidProjectId(projectIdParam) || !documentId) {
-        throw new Error(
-          'Cannot select a change since projectId or documentId are not set yet.'
-        );
-      }
-
-      return selectChange({
-        projectId: projectIdParam,
-        documentId,
-        history: history ?? versionedDocumentHistory,
-        changeId,
-        showDiffInHistoryView,
-      });
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [
-      projectIdParam,
-      documentId,
-      versionedDocumentHistory,
-      showDiffInHistoryView,
-    ]
-  );
-
-  const handleDocumentContentChange = useCallback(
-    async (doc: RichTextDocument) => {
-      if (!projectStore || !projectIdParam) {
-        throw new Error('Project store not ready yet or mismatched project.');
-      }
-
-      if (!documentId) {
-        throw new Error('Versioned document id not set yet.');
-      }
-
-      if (!versionedDocument) {
-        throw new Error('Versioned document not set yet.');
-      }
-
-      if (!representationTransformAdapter) {
-        throw new Error(
-          'No representation transform adapter found when trying to convert to Automerge'
-        );
-      }
-
-      await Effect.runPromise(
-        processDocumentChange({
-          transformToText: representationTransformAdapter.transformToText,
-          updateRichTextDocumentContent:
-            projectStore.updateRichTextDocumentContent,
-        })({
-          projectId: projectIdParam as ProjectId,
-          documentId,
-          updatedDocument: doc,
-        })
-      );
-
-      await loadHistory(projectStore)({
-        docId: documentId,
-      });
-    },
-    [
-      projectStore,
-      projectIdParam,
-      documentId,
-      representationTransformAdapter,
-      versionedDocument,
-      lastCommit,
-    ]
-  );
-
-  return (
-    <CurrentDocumentContext.Provider
-      value={{
-        versionedDocumentId: documentId,
-        versionedDocument,
-        onDocumentContentChange: handleDocumentContentChange,
-        loadingHistory,
-        versionedDocumentHistory,
-        canCommit,
-        reloadDocumentHistory,
-        onRestoreCommit: handleRestoreCommit,
-        onDiscardChanges: handleDiscardChanges,
-        commitToRestore,
-        isRestoreCommitDialogOpen,
-        isDiscardChangesDialogOpen,
-        onOpenRestoreCommitDialog: handleOpenRestoreCommitDialog,
-        onCloseRestoreCommitDialog: handleCloseRestoreCommitDialog,
-        onOpenDiscardChangesDialog: handleOpenDiscardChangesDialog,
-        onCloseDiscardChangesDialog: handleCloseDiscardChangesDialog,
-        selectedCommitIndex,
-        onSelectChange: handleSelectChange,
-      }}
-    >
-      {children}
-    </CurrentDocumentContext.Provider>
-  );
-};
