+++ src/renderer/src/app-state/current-document/index.ts
@@ -1,0 +1,3 @@
+export { CurrentDocumentContext } from './context';
+export { CurrentDocumentProvider } from './provider';
+export { type CurrentDocumentContextType } from './types';
